pyproject.toml
[project]
name = "solid-lamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "requests-mock>=1.12.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
